requirements.txt
APScheduler==3.10.1
celery==5.2.7
email_validator==2.0.0.post2
emoji==2.2.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-telegram-bot==20.3
redis==4.5.5
